use hard-coded path to doc with events
Fev have changed the format of the page that held the sheet link, and
remove the link from it to implement their own version of timezones.
Right now it's not as useful as this one. And it turns out the sheet
never changes, so we can just as well use the hard-coded one.